+++ parse/src/test/resources/ddl/alter/mysql_74.sql
@@ -1,7 +1,9 @@
-CREATE TABLE `tb_tazkjnlgrx` (
-  `col_ewzdxvnlrs` bit(1) NOT NULL DEFAULT b'0',
-  `col_trhadxfbfi` decimal(10,0) DEFAULT NULL,
-  `col_jxosxbunsc` double(179,7) DEFAULT NULL,
+CREATE TABLE `tb_tazkjnlgrx`
+(
+    `col_ewzdxvnlrs` bit(1) NOT NULL DEFAULT b'0',
+    `col_trhadxfbfi` decimal(10, 0)  DEFAULT NULL,
+    `col_jxosxbunsc` double(179, 7
+) DEFAULT NULL,
   `col_jijmncnsoa` set('enum_or_set_0','enum_or_set_1','enum_or_set_2') DEFAULT 'enum_or_set_0',
   `col_gxpgglefyh` char(40) CHARACTER SET utf8mb4 DEFAULT NULL,
   `col_pgtqyvckbi` time DEFAULT NULL,
