+++ parse/src/test/resources/ddl/alter/test_18.sql
@@ -1,15 +1,20 @@
-CREATE TABLE `tb_pnracjzmkp` (
-  `col_zoovnitdls` float(13,13),
+CREATE TABLE `tb_pnracjzmkp`
+(
+    `col_zoovnitdls` float(13, 13
+) ,
   `col_prrcfbnwew` longtext CHARACTER SET utf8mb4,
   `col_peqguxwzcy` smallint(202) NOT NULL DEFAULT '1'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 CREATE TABLE `tb_qwbjzchejk` LIKE `tb_pnracjzmkp`;
-RENAME TABLE `tb_qwbjzchejk` TO `tb_ezxnrjkvwu`, `tb_pnracjzmkp` TO `tb_jhkrixchju`;
-RENAME TABLE `tb_ezxnrjkvwu` TO `tb_nxqxynvjvx`;
+RENAME
+TABLE `tb_qwbjzchejk` TO `tb_ezxnrjkvwu`, `tb_pnracjzmkp` TO `tb_jhkrixchju`;
+RENAME
+TABLE `tb_ezxnrjkvwu` TO `tb_nxqxynvjvx`;
 DROP TABLE tb_nxqxynvjvx;
 ALTER TABLE `tb_jhkrixchju` ADD COLUMN (`col_fpmvnrnoxw` varchar(18) NOT NULL, `col_hrcqpnxnrn` text(1490675533));
 ALTER TABLE `tb_jhkrixchju` ADD COLUMN (`col_iguvbfennm` year(4) NOT NULL, `col_nhdlxtgtvs` char NULL);
-ALTER TABLE `tb_jhkrixchju` ADD `col_vyizooodsa` longblob AFTER `col_nhdlxtgtvs`;
+ALTER TABLE `tb_jhkrixchju`
+    ADD `col_vyizooodsa` longblob AFTER `col_nhdlxtgtvs`;
 ALTER TABLE `tb_jhkrixchju` ADD (`col_hwmrzghead` longtext CHARACTER SET utf8mb4, `col_adgnavqmhq` longtext CHARACTER SET utf8mb4);
 ALTER TABLE `tb_jhkrixchju` ADD (`col_adwetrnlyu` bit(41), `col_ypyzwlctxm` varbinary(86) DEFAULT '\0');
 ALTER TABLE `tb_jhkrixchju` ADD (`col_lkdcwbjkis` varbinary(61), `col_rmjdxsrzry` char(159) CHARACTER SET utf8mb4);
